refactor: move exec.go into 'internal/proc' package
